svc: Remove extra casts from UVISOR_SVC_OPCODE
We don't need the uint8_t cast. Performing a bitwise and with 0xFF is
good enough to force the range of values `id` can become. Remove the
uint8_t cast. Keep the uint16_t cast, as the opcode should be seen as a
uint16_t.

+++ core/system/inc/svc.h
@@ -99,7 +99,7 @@
 #define UVISOR_SVC_ID_UNVIC_IN         UVISOR_SVC_FIXED_TABLE(0, 0)
 
 /** Generate the SVCall opcode from the SVC ID. */
-#define UVISOR_SVC_OPCODE(id) ((uint16_t) 0xDF00 | (uint8_t) ((id) & 0xFF))
+#define UVISOR_SVC_OPCODE(id) ((uint16_t) (0xDF00 | ((id) & 0xFF)))
 
 #define UVISOR_SVC_ID_GET(target) UVISOR_SVC_CUSTOM_TABLE(offsetof(UvisorSvcTarget, target) / sizeof(uint32_t))
